LayoutTests/imported/w3c/web-platform-tests/html/webappapis/system-state-and-capabilities/the-navigator-object/plugins-and-mimetypes-expected.txt


FAIL navigator.pdfViewerEnabled exists assert_true: property exists expected true got false
FAIL navigator.mimeTypes is empty assert_equals: length expected 0 but got 14
FAIL navigator.plugins is empty assert_equals: length expected 0 but got 1
